fixed global variable implementation
