+++ GPUmodules/__init__.py
@@ -15,4 +15,5 @@
 #    You should have received a copy of the GNU General Public License
 #    along with this program.  If not, see <https://www.gnu.org/licenses/>.
 from GPUmodules import GPUmodules
+from GPUmodules import env
