+++ amdgpu-monitor
@@ -28,9 +28,9 @@ __copyright__ = "Copyright (C) 2019 RueiKe"
 __credits__ = ""
 __license__ = "GNU General Public License"
 __program_name__ = "amdgpu-monitor"
-__version__ = "v2.0.0"
+__version__ = "v2.1.0"
 __maintainer__ = "RueiKe"
-__status__ = "Stable Release"
+__status__ = "Development"
 
 import gi
 gi.require_version("Gtk", "3.0")
@@ -44,21 +44,20 @@ import platform
 import sys
 import time
 from GPUmodules import GPUmodules as GPU
+from GPUmodules import env
 from datetime import datetime
 from uuid import uuid4
 import glob 
 import shutil 
 from pathlib import Path
 
-# Instantiate project constants
-gut_const = GPU.GUT_CONST()
 
 class MonitorWindow(Gtk.Window):
     def __init__(self, gpu_list, devices):
         Gtk.Window.__init__(self, title="amdgpu-monitor")
         self.set_border_width(1)
-        if(os.path.isfile(gut_const.PATH + '/amdgpu-monitor.icon.png') == True):
-            self.set_icon_from_file(gut_const.PATH + '/amdgpu-monitor.icon.png')
+        if(os.path.isfile(env.gut_const.PATH + '/amdgpu-monitor.icon.png') == True):
+            self.set_icon_from_file(env.gut_const.PATH + '/amdgpu-monitor.icon.png')
         grid = Gtk.Grid()
         grid.override_background_color(Gtk.StateType.NORMAL, Gdk.RGBA(1,1,1,1))
         self.add(grid)
@@ -149,21 +148,21 @@ def main():
         print("Status: ", __status__)
         sys.exit(0)
 
-    gut_const.PATH = os.path.dirname(str(Path(__file__).resolve()))
-    if args.debug == True :  gut_const.DEBUG = True
+    env.gut_const.PATH = os.path.dirname(str(Path(__file__).resolve()))
+    if args.debug == True :  env.gut_const.DEBUG = True
     if int(args.sleep) > 0 :
-        gut_const.SLEEP = int(args.sleep)
+        env.gut_const.SLEEP = int(args.sleep)
     else:
         print("Invalid value for sleep specified.  Must be an integer great than zero")
         sys.exit(-1)
 
-    if gut_const.check_env() < 0:
+    if env.gut_const.check_env() < 0:
         print("Error in environment. Exiting...")
         sys.exit(-1)
 
     # Check value of AMD Feature mask
     try:
-        featuremask = gut_const.read_amdfeaturemask()
+        featuremask = env.gut_const.read_amdfeaturemask()
     except FileNotFoundError:
         print("Cannot read ppfeaturemask. Exiting...")
         sys.exit(-1)
@@ -173,7 +172,7 @@ def main():
         print("AMD Wattman features not enabled: %s, See README file." % hex(featuremask))
         sys.exit(-1)
 
-    gut_const.get_amd_driver_version()
+    env.gut_const.get_amd_driver_version()
 
     gpu_list = GPU.GPU_LIST()
     gpu_list.get_gpu_list()
@@ -204,7 +203,7 @@ def main():
         gmonitor.show_all()
 
         # Start thread to update Monitor
-        monthread = threading.Thread(target=refresh, daemon=True, args=[gut_const.SLEEP, updateData, gpu_list, devices]).start()
+        monthread = threading.Thread(target=refresh, daemon=True, args=[env.gut_const.SLEEP, updateData, gpu_list, devices]).start()
 
         Gtk.main()
     else:
@@ -213,9 +212,9 @@ def main():
             while True:
                 gpu_list.read_hw_data()
                 gpu_list.read_device_data()
-                if gut_const.DEBUG == False: os.system('clear')
+                if env.gut_const.DEBUG == False: os.system('clear')
                 gpu_list.print_table()
-                time.sleep(gut_const.SLEEP)
+                time.sleep(env.gut_const.SLEEP)
         except KeyboardInterrupt:
             sys.exit(0)
