+++ amdgpu-pac
@@ -28,9 +28,9 @@ __copyright__ = "Copyright (C) 2019 RueiKe"
 __credits__ = ""
 __license__ = "GNU General Public License"
 __program_name__ = "amdgpu-pac"
-__version__ = "v2.0.0"
+__version__ = "v2.1.0"
 __maintainer__ = "RueiKe"
-__status__ = "amdgpu-pac"
+__status__ = "Development"
 
 import gi
 gi.require_version("Gtk", "3.0")
@@ -44,21 +44,20 @@ import platform
 import sys
 import time
 from GPUmodules import GPUmodules as GPU
+from GPUmodules import env
 from datetime import datetime
 from uuid import uuid4
 import glob 
 import shutil 
 from pathlib import Path
 
-# Instantiate project constants
-gut_const = GPU.GUT_CONST()
 
 class PACWindow(Gtk.Window):
     def __init__(self, gpu_list, devices):
         Gtk.Window.__init__(self, title="amdgpu-pac")
         self.set_border_width(1)
-        if(os.path.isfile(gut_const.PATH + '/amdgpu-pac.icon.png') == True):
-            self.set_icon_from_file(gut_const.PATH + '/amdgpu-pac.icon.png')
+        if(os.path.isfile(env.gut_const.PATH + '/amdgpu-pac.icon.png') == True):
+            self.set_icon_from_file(env.gut_const.PATH + '/amdgpu-pac.icon.png')
         grid = Gtk.Grid()
         grid.override_background_color(Gtk.StateType.NORMAL, Gdk.RGBA(1,1,1,1))
         self.add(grid)
@@ -581,7 +580,7 @@ class PACWindow(Gtk.Window):
         fileptr.close()
         os.chmod(out_filename, 0o744)
         print("Batch file completed: %s" % out_filename)
-        if gut_const.execute_pac == True:
+        if env.gut_const.execute_pac == True:
             print("Writing changes to GPU %s" % v.card_path)
             cmd = subprocess.Popen(out_filename, shell=True)
             cmd.wait()
@@ -650,7 +649,7 @@ class PACWindow(Gtk.Window):
         fileptr.close()
         os.chmod(out_filename, 0o744)
         print("Batch file completed: %s" % out_filename)
-        if gut_const.execute_pac == True:
+        if env.gut_const.execute_pac == True:
             print("Writing changes to GPU %s" % v.card_path)
             cmd = subprocess.Popen(out_filename, shell=True)
             cmd.wait()
@@ -692,17 +691,17 @@ def main():
         print("Status: ", __status__)
         sys.exit(0)
 
-    gut_const.PATH = os.path.dirname(str(Path(__file__).resolve()))
-    if args.debug == True :  gut_const.DEBUG = True
-    gut_const.execute_pac = args.execute_pac
+    env.gut_const.PATH = os.path.dirname(str(Path(__file__).resolve()))
+    if args.debug == True :  env.gut_const.DEBUG = True
+    env.gut_const.execute_pac = args.execute_pac
 
-    if gut_const.check_env() < 0:
+    if env.gut_const.check_env() < 0:
         print("Error in environment. Exiting...")
         sys.exit(-1)
 
     # Check value of AMD Feature mask
     try:
-        featuremask = gut_const.read_amdfeaturemask()
+        featuremask = env.gut_const.read_amdfeaturemask()
     except FileNotFoundError:
         print("Cannot read ppfeaturemask. Exiting...")
         sys.exit(-1)
@@ -712,7 +711,7 @@ def main():
         print("AMD Wattman features not enabled: %s, See README file." % hex(featuremask))
         sys.exit(-1)
 
-    gut_const.get_amd_driver_version()
+    env.gut_const.get_amd_driver_version()
 
     gpu_list = GPU.GPU_LIST()
     gpu_list.get_gpu_list()
